Clear the IR1 namespace after component compilation.
Not only does this reduce garbage retention for large compilations,
but doing it this way helps ensure that we don't have any references
to IR1 referring to stuff in an already compiled (i.e. "dead")
component. We no longer need special logic just to clear specific
outdated stuff (as in CLEAR-INVALID-FUNCTIONALS).

Originally we must've done this, as this logic is present in CMU CL,
but at some point clearing the IR1 namespace got removed and hence the
invalid functional hack got added. I can't ascertain why the IR1
namespace clearing was removed in the first place.