Speed up pinned_p
Move 'pinned' out the page table, and use 1 byte per page instead of
1 bit per page.  Zeroing all the pinned bits is much faster now because
memset() outperforms reading and writing one byte per PTE.

Having increased the number of pin tracking bits, refine the test which
bypasses the pinned_objects hashtable. Map an object to a subrange of
its page and check whether 1 bit in the byte is set for that subrange.
This seems to reduce the number of hopscotch_containsp calls to about 1%
of all uses of pinned_p() where formerly it was over 2%.
Naively I expected an 8x reduction, but it's an improvement.

Other changes:
- clear the bits only once per collect_garbage() instead of making a
  pass over them for each collect_generation().
- with LRAs, do less work in pin_object() and always translate the LRA
  to a codeblob address when asking if the LRA if pinned.