Separate out the compile time only side effects of PROCLAIM.
Some aspects of PROCLAIM only happen at compile-time. This is a bit
anomalous because ANSI functions don't usually cause compile-time side
effects, instead using a macro. That is the reason of being for
DECLAIM, but we currently take the position that PROCLAIM happening at
compile time need to cause the same effects, so separate
out (textually) some of the compile time only effects to make it clear
that is what is happening.