Make it more clear why we care about DEFINED-FUNs from NAMED-LAMBDA.
More generally, we'd like to distinguish the case of a defined fun
coming from the same compilation block -- only in that case do we
actually want references to resolve to the functional defined in the
same block. That still preserves object identity from the
reference. Otherwise, we need the reference to refer to the global
function, since the function was defined outside the block and we
don't want to reference a new entry point for that function.
RECOGNIZE-KNOWN-CALL is in charge of actually doing the resolution
specifically in the case of a function call, rather than a
reference. In that case it doesn't matter whether the defined-fun
comes from the same compilation block or not.