Streamline free_oldspace
Delete detection of contiguous block boundaries. The only purpose that
could have served was to visit objects in the freed block. But you can't-
see cautionary note about how not to compute an object size.
In light of that, I think the only way to implement visit_freed_objects()
would be if forwarding headers stored the original object size.