Make block compilation of top-level closures work.
Broken since at least SBCL prehistory, since we removed some of the
machinery that lets this stuff work correctly. Now we do essentially
what cmu cl does, except that we improve upon that by allowing lexical
environments with FLET. Really instead of caring about the lexical
environment we should check that the runtime environments coming from
environment analysis are compatible, to optimize this to the fullest
extent. But at least we can now block compile and correctly execute
the code at all.

Also, in IR1-convert-inline-expansion, we used to test for
EXPANSION. This was wrong ever since we stopped passing the expansion
from INLINE-SYNTACTIC-CLOSURE-LAMBDA through, because the (WHEN
EXPANSION ...) was really just testing for a hairy lexical
environment. Use the new SIMPLE-LEXICAL-ENVIRONMENT-P which ignores
local functions instead.

Scrubbing forward references actually has semantics for block
compilation now, so move that stuff out of the proclamation machinery
and back into ir1tran.