Rename OBJECT_SIZE and friends
There are 4 sizing functions now, all the same but different:
 OBJECT_SIZE() -> object_size2() // meaning it takes 2 args
 with special case of OBJECT_SIZE(*x,x) as object_size(x)
and
 sizetab[widetag_of(x)](x) becomes headerobj_size(x) etc

The concept is that object_size() takes any object but headerobj
only accepts headered objects. It may be wishful thinking
that selection of the right variant removes some branching,
but they can be regarded as assertions about the logic.