Refactor filesys tests using new macro WITH-TEST-DIRECTORY
Regrets about having tests write into the source directory are now
centralised into this new utility.

In addition to abstracting the existing pattern of creating then
cleaning up a test directory based on the TEST_DIRECTORY environment
variable, WITH-TEST-DIRECTORY also interacts with CHECK-MANIFEST to
avoid hardcoding the names of temporary files within the test
directory. (a.txt and b.lisp in particular are only being touched and
could be opened with :DIRECTION :OUTPUT instead to avoid tripping on
CHECK-MANIFEST. But this behaviour will be useful in a future test.)