Fix extremely bothersome semantics of zero_dirty_pages()
There were two problems with this function:
1. The implication of the name is that would leave each page with
   need_to_zero as false after clearing, but it actually marked the
   zeroed pages as needing zeroing. So just on the surface of it,
   that's highly counterintuitive.
2. Recently it acquired a new aspect of not doing anything on
   certain pages, because GC-use pages (as opposed to mutator-use pages)
   can generally be utilized without ever prezeroing them.

Had the dubious function been instead named "prepare_page_for_use"
it might have at least not been a complete lie. But both problems
can be fixed while making it more efficient too. Need_to_zero is
relevant only if the page is free, and about to become used, and had
been used before. Therefore, only when freeing from_space pages must
it be recorded that a page was ever dirty. And when shrinking an object
which is basically the same as freeing from_space.

The revised behavior confers a slight advantage in this scenario:
 - suppose 3 consecutive pages start as clean (need_to_zero == 0)
 - a new region spans those pages, then we allocate to the first page.
 - close the region.
 - two pages are still untouched, so their need_to_zero stays as 0.
 - reset_page_flags on the last 2 pages: still not-need-to-zero.
This really happens. The old behavior would have put all 3 pages in the
newly opened region into need_to_zero state simply because they became
part of a region.