Fix another defstruct problem
I think this problem is new as of this month but I'm less sure about it
than with the spurious redef warning. If the compiler knew about a structure
type, as it must have in order to dump a literal of the type, then it must
have processed a %DEFSTRUCT, but possibly not caused the effect of that
defstruct to happen at load-time by the time that FOP-STRUCT is invoked.