Work around remaining RENAME-PACKAGE issue for tll smashing.
Simply force compilation of queued lambdas when a
DEFPACKAGE/MAKE-PACKAGE is seen at compile time, thereby avoiding
issues with RENAME-PACKAGE. Set TOP-LEVEL-LAMBDA-MAX back to 20. This
is for the case which I think is not conforming code, but nevertheless
causes behavioral change. It's not worth trying to solve this to save
a few bytes off of fasls in the case where we have random package
hackery going on, something which doesn't apply to the majority of
code.