genesis: Create cold packages "on demand".
By creating packages on demand only for those symbols that actually
get loaded during genesis, we no longer create packages in the target
that aren't needed. This makes cross-compilation and cold loading
custom builds with stuff added or removed more flexible. It also
allows us to create packages normally without the extra indirection of
making package data objects in the future. This also centralizes the
creation of the packages COMMON-LISP, COMMON-LISP-USER, and KEYWORD,
allowing docstrings to be initialized all in one place.

We also define a cross compiler version of PACKAGE-NAME that returns
COMMON-LISP for the XC-STRICT-CL and SB-XC packages, since this could
potentially get confusing.