Add new timing wrapper for parallelized test suite
parallel-exec uses sb-aprof and sb-sprof and collects data about
vops used, and injects extra GCs, which is not a a fair test of
whether code generation got better; and if the compiler gets
slower, it also makes the test suite take longer. Those facts made
it difficult to get a sense of whether the suite is executing faster.
To this end, 'benchmark.sh' does the same thing as parallel-exec,
but less of it, and also subtracts time spent in the compiler.
Take out the GC that I recently added to PURE-RUNNER, same reason.