Remove REEXPORT clause from PACKAGE-DATA.
We needed them since the beginning of SBCL because EXPORT clauses
inside package data were just not handled in the right order. But now
they are, so we don't need to use :REEXPORT most of the time, we can
just use normal DEFPACKGE machinery. For the cases where REEXPORT was
just an abbreviation for IMPORT-FROM, then EXPORT, we can just write
out the exports again manually. If we don't want to repeat this we
could also define a function REEXPORT to do this. The idea is that we
want the package definitions forms used to set up cold packages truly
look the same as normal DEFPACKAGE. Also massively simplifies the
actual creation of packages.

This also makes SB-C:SLOT a different symbol from SB-ALIEN:SLOT. It
didn't make sense to conflate them given they don't share any
commonality in terms of semantics or definitions.