Try to avoid SB-DI::*FUN-END-COOKIES*
Tried to fix it on ppc but failed, presumably due to the instructions
around the comment saying
"I have literally no idea why this constant is -24" in ppc-assem.S.
Some of the others have MAKE-BPT-LRA signaling an error.
So at least a few architectures pass the tests in debug.impure
with the modified code.