mips: change one kludge to another
Non-use of pseudostatic pages in core.h was a truly inexplicable hack.
I found a bug ever since rev 89fb96640ab0 reduced GENCGC-PAGE-BYTES to
smaller than +BACKEND-PAGE-BYTES+. Apparently we never anticipated those
parameters diverging on any OS where we load the core file to memory
using mmap(). Perhaps coincidentally, win32 is the only parms file that
defines the two unequally, and it does NOT use MapViewOfFile.
I wouldn't be surprised if we only thought it didn't work because of this.
(Which came first: avoidance of MapViewOfFile or divergence of parameters?)