Simplify fullcgc
Don't mark objects in their headers. It's faster to use a side table.
This is partly because of special logic necessary to mark simple-funs
(marking a function marks all others in the same codeblob), but it's
also true even without that aspect. We don't have to revisit all
objects to clear the header bits on completion of the sweep phase,
and there's no longer anything different about cons cells.