Add some flexibility to the build
Do not assume that 'cc' is executed in src/runtime. In particular,
allow the root of SBCL's source tree to be one or more levels down
from where the compiler gets invoked. This places different demands
on where autogenerated headers are written to by 1st genesis and
where they are read from by make-target-1 (or equivalent).

And non-system includes should not use <> syntax (in main.c)