Remove non-pauseless-threadstart code
This is a modified version of Douglas Katzman's commit 869f2acd6217
(Remove non-pauseless-threadstart code), which gets the
pauseless-threadstart feature to work on Windows.  Since Windows is the
last outstanding platform that builds with pauseless-threadstart
disabled by default, this commit also removes the pauseless-threadstart
feature flag entirely.

The original commit was modified so that it applies to 2.2.7 as well as
to remove the setup semaphore and to set th->no_tls_value_marker to
NO_TLS_VALUE_MARKER in alloc_thread_struct.

Note: The original commit was reverted by 15f2ccc55127 (Revert "Remove
non-pauseless-threadstart code") due to failing tests.