Mostly forbid duplicate DEFUNs in genesis
I don't know why rev 3bdd4d28ed took out the lines
    (unless (cold-null (cold-fdefn-fun fdefn))
      (error "Duplicate DEFUN for ~S" warm-name))
from COLD-FSET, but it did, and of course in the interim there
was backsliding such that there is now a duplicate.

And lexical vars named DEFN and FDEFN are too confusing to
co-exist in the same function, so fix that.