Remove "augmented" symbols
Now that :function :definition is not used by globaldb for fdefns of
symbols, we can repurpose that to store the special-form handler for
fast-eval instead of extending special operators by one slot.
And there were no benchmarks showing that augmented symbols were useful
to begin with, so this just as good, and with less complexity.