Don't check for VECTOR before calling hairy-data-vector-ref.
Because it dispatches on the widetag. Since hairy-data-vector-ref can
actually work on any array it has to be hard-coded in
cast-externally-checkable-p. The error message already says "not of
type vector", which is not exactly right for arrays, but in safe code
a multidimensional array will be checked when calculating the row
major index.