Speedup float-math verification
Because it was consing ~400MiB and taking nearly a second, I was inclined
to perform it only if #+sb-devel. Then I realized the problem: EVAL is not
operational yet, therefore "#." was effectively equivalent to COMPILE.
So reading compiled nearly 8000 forms before even starting to work on
any files. Instead intercept "#." to do something ad hoc and fast.
It conses almost nothing now, except the floating-point values.