x86-64: Rework assembly routine fixups
Make a vector of entrypoints in static space so that dynamic-space code can
use "CALL [ea]". We don't need to retain the fixup for reapplication if the
text space moves on core start. Removes some complexity from rev 7d84b420e0
("create fewer absolute fixups") but keeps the level of indirection for
allocation profiler fixups, because that logic is finicky and needs the CALL
to encode in 5 bytes. This gets a little nearer to being able to relocate
text space to any address, not just sub-2GB.