win32: handle and recover from stack overflow exceptions
Create threads with STACK_SIZE_PARAM_IS_A_RESERVATION, similarly to
what happens with the main thread. (This explains why up until now it
was possible to handle a stack overflow exception, albeit only once,
on the main thread but not other threads.)

With this option, stack memory is reserved rather than committed. This
has the benefit of using less memory upfront and also lets windows set
up a guard page (actually a region with several pages) that takes care
of committing more pages as the stack grows or raising a
STACK_OVERFLOW_EXCEPTION.

Use SetThreadStackGuarantee() to customise how much stack space should
be reserved for handling stack overflow exceptions. This is similar to
what SBCL does with CONTROL_STACK_SOFT_GUARD_PAGE on other platforms.

The last page in the stack is always protected and Windows will
terminate the process if it's triggered. (Similarly to what we do with
the CONTROL_STACK_HARD_GUARD_PAGE on other platforms.)

Instead of protecting a CONTROL_STACK_RETURN_GUARD_PAGE immediately
when a stack overflow exception is raised, we protect a region below
the stack guarantee upon unwinding so that Windows may manage the
stack the next time we attempt to overflow it.

Summing up:

  1. It is now possible to handle stack overflows on any thread rather
     than just the main thread.
  2. We can handle and recover from stack overflows multiple times.
  3. There's enough stack available to be able to, e.g., debug a stack
     overflow in SLIME.
  4. (Bonus points) Stack overflow recovery is now slightly more
     resilient on Windows compared to other platforms. See new
     tests (:EXHAUST :WRITE-TO-STACK-ON-UNWIND) and (:EXHAUST
     :WRITE-TO-STACK-IN-HANDLER).

Change-Id: I62bb1753dd7e04c40e7f14f2a5e9bb0e00ec0fc7