Add some declarations for allocator control
and a design doc explaining why.

Unless compiled in with #+system-tlabs, there is no change to the
generated code. So average users won't care either way.

The implementation of this feature is much further along than implied
by the design. Which is to say, it works, and will be committed
in a series of changes, pending some touch-ups.