Add a feature to fullcgc for locating unexpected pointers
A user-supplied function can be specified which performs any test
you like on all pointers that are NOT of interest to the trace
but are of interest to you.
e.g. you can find pointers to the stack or alien memory.
It doesn't matter if this slows fullcgc down any.