Remove SB-PCL::STANDARD-BOUNDP-METHOD
STANDARD-READER-METHOD and STANDARD-WRITER-METHOD are classes used to
represent methods added by :reader/:writer/:accessor arguments
in (direct) slot specifications; having this class identity can be
helpful as it allows the system to keep track of whether *all* methods
of a generic function are readers (or writers), and if so to
short-circuit some of the more general CLOS dispatch and effective
method computation.

There is no interface to represent methods which perform boundp (or
makunbound) operations directly on slots; no :boundp argument (for
example) in a slot specification.  SB-PCL::STANDARD-BOUNDP-METHOD has
been in the system since SBCL's initial revision, and was probably
added in PCL to be analogous to READER/WRITER, but the analogy is
broken given the absence of such a way of specifying names of
functions in slot specifications.  There does not seem to be a clamour
for such functionality to be added, and no-one to my knowledge has
ever created a STANDARD-BOUNDP-METHOD; so remove it.