Implement a general hashset, use it for pathname cache
This change resulted in a huge time saving in INTERN-PATHNAME. Tested on
the same operations, the former hash algorithm required as many as 120 probes
worst case, where this hashset needed at most 8 probes, with the mean
being between 2 and 3 probes.

Also, it's threadsafe for reading (but not truly concurrent- just not
unsafe), so it does not use a mutex for lookups except if not found,
in which case it retries with a lock and possibly insertion.
The memory use is typically under half that of a hash-table having the
same keys, and it also measurably outperforms a weak hash-table.