Remove DX + CAST interaction "workaround".
Probably conceptually broken for the past decade after environment
analysis was made to do deep checks as well. We need to protect the dx
lvars generated from CAST in the cleanup as well, since they need to
participate in stack analysis. No problems had arisen because we were
making a new basic block to workaround the problem and have stack
analysis not scan the block as a pusher. But this unnecessarily
creates basic blocks just for CAST nodes when we should be protecting
all the dx lvars anyway. The comment stating that dx value generators
need to end their blocks for the sake of stack analysis was outdated
too (it was only necessary to work around the aforementioned
problem). So just do the right thing.

The test case added wasn't actually triggered before, but it showcases
a scenario where you can't simply ignore dx-lvars generated from CAST
nodes entirely.