Change the cutover heuristic for [N]UNION
The decision to use a hash-table was based entirely on the shorter list,
and the threshold for switching was too high. In testing this change
I tried a variety of inputs and almost always realized a 2x to 3x speedup
by choosing a hash-table more often. Not to mention, there's an easy case
that is a zillion times faster: when one input is NIL, don't call LENGTH.
There are cases where the old heuristic did better, but it's impossible
to predict because it depends on how much the inputs have in common.