Make DEFCONSTANT-EQX use DEFVAR on the host
DEFCONSTANT-EQX is used, in general, for structured data.  If it is
defined as a true constant, then it is implementation-dependent
whether that structured data gets effectively inlined by the compiler
into the cross-compiler data structures, which then implies that the
cross-compiler can have differences in EQLity of its data structures,
which can emerge as a reproducibility issue for the target fasls.

This fixes the reproducibility of simd-pack backend files on x86-64.