Slightly more helpful SHOW-FOP printing for fop-push
It's possible to reconstruct what's going on from looking at the fops
at the creation site of the table entry, but printing what we can find
there is a helping hand (even if it's not as helpful as it might be
when looking at cold-fops in genesis, where many things get printed as
unreadable #<DESCRIPTOR ...>).

It might also be possible to make the output more generally useful
for fops, inspecting not only the fop operands but also the input
stream for things like fop-byte-integer.  This would probably entail
making the fop definitions more declarative and is left as an exercise
to the ambitious maintainer.