stack: Don't scan each block's nodes during pathwise dx propagation.
We simply do the propagation while zooming up the cleanup nesting and
checking if the current block is the mess-up block. This requires
inserting a %CLEANUP-POINT funny function like we already do for DX
closures.

THis also sort of removes an assumption on what the type of the
mess-up node has to be for DX cleanups.