Increase density of package symbol vectors
PACKAGE-HASHTABLE-IMAGE-LOAD-FACTOR being less than the rehash threshold
made no sense whatsoever, as it caused all packages to use _more_ memory
than they would have prior to "tuning" them for dump. It would have been
ok if it were the stated objective (perhaps to improve the symbol hit rate)
but the comment said specifically that the goal was to optimize for space.

So do exactly what the comment says- decrease the space used, and moreover
apply the Robinhood strategy when (and only when) rehashing, and adjust the
desired load factor depending on which package it is.

Also: rename PACKAGE-HASHTABLE to SYMBOL-HASHSET. It was confusing
that some comments mentioned a "package hashtable" in reference to
the global hashtable mapping names to packages.