Implement fastrem-32 vop for arm64 and ppc64
This got about 5% speedup in the FIND-SYMBOL benchmark for those CPUs,
less than on x86-64, but I'll bet they run the string hashing step more
slowly, so the remainder operation in %LOOKUP-SYMBOL is a smaller
contributor to the total time. Amdahl's law pertains.

Also implement fastrem-64 in case we need it.