Use obj/from-self as the location of all warm-target-compiled fasls
There was no such directory when I changed warm.lisp to not leave fasls in
the source tree (rev 336d4be26e).  But contribs go there now, and it's
silly to put a defvar in target-load to specify this path, so keep all files
compiled by the post-bootstrap image together in a place that makes sense.