Fix subtle bug in auto-defined ARRAY-TYPE-EQUIV
At least one regression tests contradicts an assumption made in the hashset.
The (:hasher NIL) option on specialized-element-type says that not only does
it not need to be mixed in to obtain a hash, it may also be disregarded in the
equivalence predicate. Based on that I would have expected the hashset to
return an instance that did not match the constructor parameters. But no test
failed; I discovered this completely by accident.