Delete most special cases from CTYPE constructors
The extra testing entailed by choosing to return constant objects
is for nothing, since hash-consing achieves the same thing.
MAKE-NUMERIC-TYPE in particular is now much more understandable.

Remove the 5-bit type-saetp-index field embedded within the
type-hash slot, and allow those bits to be random.

Remove *PARSED-SPECIALIZED-ARRAY-ELEMENT-TYPES* which is
redundant with the SAETP-CTYPE of each SAETP instance.