Only run dead code elimination once,
DCE was keeping lambdas alive even if they had cross-component
references which themselves had already been condemned to
deletion. This was causing some accidentally-kept-alive dead code to
puke on itself, which is why commit
ec52ae844db3f9739adb571ca8739052e37ff203 introduced running DCE during
initial DFO and then _also_ repeatedly during optimization as a
workaround, But DCE is a full blown flow analysis so it's
better/faster if we just run it once instead of repeatedly all the
time, particularly because the time it takes is proportional to the
amount of live code rather than the amount of dead code to be deleted,
and code to be deleted is less common than code not to be deleted. By
fixing the actual bug, we can get rid of the extra calls to DCE which
were just a workaround for the true problem.