Fix some remainng package weirdness.
Get rid of deferred packages and so on. This makes two tests fail
under block compilation mode, but fixes a very real correctness
problem in normal evaluation mode. A serious rethink needs to be done
to figure out how to deal with fasloader and package interaction under
block compilation. At least we solved the accessibility issue
correctly with the use of a new loader op. Now the package mechanisms
are back to what they looked like before the implementation of
deferred packages and should be at least as correct as the state of
affairs before that was added.

Fixes lp#2000004.