Remove rehoming packages.
Should solve lp#2000004 this time without regressing the compile-twice
test. Trades the correctness of a package test under block
compilation, but this is more important to get right, since we were
doing stuff with symbols that weren't even manifest constants in
compiled code, so we have no right to treat them wonkily. The better
thing to do is to evict symbols that have only been interned in the
loader.