Unify NLX tag checking analysis with indirect variable analysis.
Because they are really doing the same thing. Use the existing
variable analysis for both because it is more obviously correct and
also more straightforward and much more efficient. The old tag
existence check analysis was doing a recursive search for "escape
paths" using a hash table to break cycles. It's much easier to just
scan the closure environments with possible external entry points,
since the extent of the home environment relative to where the
variable is used/closed over is what we really care about. The upshot
is we can analyze both indirect lambda vars and nlx infos in one go.

Since the analysis is more obviously correct now, do cell elision even
on SAFETY 3.

We also needed to change the representation of NLX-INFO to facilitate
grabbing the EXIT out of the NLX-INFO. There is still some redundancy
left over from when continuations got split into lvars and ctrans that
still need to be squeezed out.