Remove randomizing effect of memoizing UNKNOWN-TYPE anywhere
This corrects a very ancient issue: the better you cache, the less you evict;
the less you evict, the more UNKNOWNs pollute the cache, the more unknowns
cached, the more the compiler gets screwed up.
So the compiler has always depended _critically_ on cache line collisions.
I was able to produce a miscompile of SBCL under SBCL by randomly scrambling
the order of some DEFUNs. There's no way to know which type operations come out
wrong except by hand-examining a bunch of simple-fun types in the pristine core
based on a-priori knowledge of the right answer. Internally, intermediate
results could come out wrong, but it's hard to write a self-check because it's
a turtles-all-the-way-down scenario. (asking some other function if it's wrong,
when it's wrong about whether it's wrong)

This is a minimal fix that calls CONTAINS-UNKNOWN-P too often. It can/should be
computed exactly once per instance when inserting to a hashset,
and stored by taking 1 bit from the %BITS slot.

I still do not know how there was a dependence on the cross-compilation host,
but this seemed to fix the fasl discrepancy with CCL, so I guess CL:SXHASH
is used in a way that affects hash-caches.

Two regression tests are broken on purpose because they're not as important
as the problem that this change solves.