Remove the CTYPE-OF cache
It pre-dates the ability to return literal ctypes given a constant call
to SPECIFIER-TYPE, and pre-dates hash-consing in general.
So it doesn't do much. Plus there is no such cache in make-host-1.
Or "what's good for the goose is good for the gander"
where "goose" = make-host-1 and "gander" = everything else.