Add some inline declamations, improve some type declarations.
We should prefer declaring argument and values types in system code
with DECLARE rather than function ftype proclamations, since the type
checkiung behavior is actually different, and it is more difficult to
maintain the ftypes. Also, the reason why LABEL-POSITION was
miscompiled was probably because of the declaration on the optional
argument (which naturally can be NIL). Switching to a declaration
syntax makes that obvious (and the compiler actually catches it that
way).