Use notice_pointer_store() correctly
With one exception, soft card marking normally requires an object header's
card to be marked, and *not* necesarily the card with an old->young pointer.
(The exception being that SIMPLE-VECTOR marks element-by-element)

So this code was wrong but I think it worked anyway for hash-tables, because
they are relocated to PAGE_TYPE_SMALL_MIXED where card-spanning is impossible.