Adjust code to match comment
The comment quite correctly said that the simplification in the union
of two cons types only applied if the types of the cars was not
disjoint.  This was not being correctly tested (in fact I'm struggling
to reverse-engineer what *was* being tested) and so we could proceed
to simplify to the same type, while claiming to have made progress,
leading to a stack overflow in the type system.

Resolve this by explicitly testing the intersection, comparing it
against the empty type for disjointness; fixes lp#1999352.