Fix extremely broken definition of search_dynamic_space()
I don't know how this was working as reliably as it did. It "should have" crashed
(and does crash) on any page with a random bit pattern below the argument pointer
but above the page's current usage. How could that happen? Simply: it was not
restricting the search below the page_bytes_used. It relied on extreme good luck
not to encounter a bad bit pattern before finding an object enclosing the potential
pointer.  This is a _terrible_ strategy. Consider: any page that last held code can
be assumed to have random junk everywhere, and pages that GC copies objects to do
not get prefilled with a safe bit pattern.

The down-side of enforcing the limit is that it will never find things in an open
region. That's the lesser evil, because the greater evil is to crash during
stack scanning (any of the variations) in collect_generation as follows:
  maybe_pin_code -> dynamic_space_code_from_pc -> search_dynamic_space
   -> gc_search_space -> "no object sizer for xxxx"