Remove ASSERT-GLOBAL-FUNCTION-DEFINITION-TYPE, fixing 2 FIXMEs.
As alluded to by its comment, it's basically doing exactly what
ASSERT-NEW-DEFINITION is doing, just without the enlightening
comment. In fact, ASSERT-NEW-DEFINITION was always there and was
originally used before the IR1 %DEFUN translator stuff got rewritten,
causing its use by DEFUN to be wrongly removed. That caused a type
checking regression which was only fixed much later by adding
ASSERT-GLOBAL-FUNCTION-DEFINITION, without it being realized that
ASSERT-NEW-DEFINITION already exists and does what is needed to check
proclaimed types.

So just use ASSERT-NEW-DEFINITION again instead, also fixing the FIXME
saying the name didn't quite match up with its use in
IR1-CONVERT-INLINE-EXPANSION. The reason is now clear: it was supposed
to be used in the other places too, not just for the inline expansion
case. We are getting closer to reunifying the inline expansion
conversion case with normal defun handling as was done before the IR1
%DEFUN translator rewrite.