Simplify processing inline lambdas and lexenvs even more.
Commit ad792784dac16d1d4de03afa72cb4e145e6d15bf introduced
PROCESS-INLINE-LEXENV because nested macrolets could cause things to
go haywire when inlining lambdas with lexenvs. But now that we
macroexpand away the macro environment, this is no longer an issue. We
don't even have symbol macros or macros to close over anymore, so we
don;t need to call out to compile either. Declaration nesting is not
an issue either so we can just build one flat lexenv which is way
simpler.

Also remove an unused keyword argument to IR1-CONVERT-INLINE-LAMBDA.

Incidentally, make sb-fasteval also macroexpand-all, to make it
conform to this much simpler model of doing things.

Also fix a bogus :allow-style-warning and comment in
alien.impure.lisp. The explanation made no sense and the issue was
probably that the interpreter was doing something wrong when creating
a LAMBDA-WITH-LEXENV. It works now.