Implement the global package table differently
As much as it would be nice to use a standard SBCL hash-table,
the basic hash-table is not threadsafe for reading without a lock.
The code in src/code/target-lfhash doesn't implement deletion
which is needed for DELETE-PACKAGE and doesn't allow for 0-filling
the old storage vector when growing it.
This change implements a custom table that does implement deletion
and 0-fill, removes the need for calling %rebuild-package-names
after deletions, and is comparable in speed to the code in lfhash.