Simplify IR2 conversion of closures and improve commenting.
Commit 54e4fec64ac8a3be95d163c8fa441255b6df4351 removed an unreachable
case but missed simplifying the local function PREPARE which can now
be simply inlined. Upon further observation we also see that the
comment pertaining to top-level closure pre-analysis should belong to
IR2-CONVERT-ENCLOSE instead (since that deals with closure
preallocation. CMU CL did not separate the action of closure
allocationa and closure reference which is why the comment was in the
wrong place). We see from that that the LAMBDA case is pretty much the
same as the LAMBDA-VAR case: we don't actually need to substitute in a
NIL constant; we can just leave it uninitialized in that case. We
incidentally shuffle around some conditional tests there to make
things slightly faster.