Check actually live variables rather than the lexical environment.
We want to mark variables that are in the closure or variables
available as arguments in the stack frame as debug live, which we can
directly find by scanning the lambda's variables and its closure. This
is conceptually what is being done and is also faster than doing a
double nested loop over the lexical environment, which is just an
approximation for this property anyway.