Simplify and fix a DX propagation bug.
The function

(lambda ()
  (declare (optimize debug))
  (let ((x (let ((z (make-array 3)))
             (let ((x z))
               (unknown x))
             (setf (aref z 0) 10)
             (setf (aref z 0) 20)
             z)))
    (declare (dynamic-extent x))
    (aref x 0)))

incorrectly stack allocated Z. This is probably because the logic
surrounding dx propagation through LETs has become imprenatrable. Try
to rectify this by simplifying the code a bit and then fixing the
issue. Somehow 9 lines of code were written to check whether a LAMBDA
is a let (by checking to see if there's just one REF of the
lambda). Instead, we can just check to see if the functional kind is
:LET, which is more robust anyway. It's also not possible for an lvar
to be EQ to its TRIVIAL-LAMBDA-VAR-REF-LVAR, since the use of the
former is a REF to a LAMBDA-VAR which is being bound to the value of
the TRIVIAL-LAMBDA-VAR-REF-LVAR, so there's no way they could be the
same. Then we can fix the bug by noting that we actually want
REF-GOOD-FOR-DX-P in the local case to check for ALL references being
okay, because we are propagating "forwards" from a bound value, not
"backwards" from a returned value. This fixes the issue.