Remove PRINCIPAL-LVAR-DEST and PRINCIPLE-LVAR-DEST-AND-LVAR.
A function already existed for a long time way before the invention of
these to do exactly the same thing - PRINCIPAL-LVAR-END does exactly
what PRINCIPAL-LVAR-DEST-AND-LVAR does and PRINCIPAL-LVAR-DEST is
merely its primary value. So use the original function.

Maybe we wouldn't be reinventing the exact same function over and over
again if we described/commented them or at least named them so its
obvious what they do. Anything involving casts and cast chains is a
particular offender in this regard.