Significantly simplify HANDLER-BIND expansion.
Now that we have anonymous closure dx propagation working, we don't
need to rewrite lambdas into DX-FLETs manually in the
macroexpander. However, we still do need to recognize having a lambda
expression as a handler to elide argument checks, which is a bit sad,
and doesn't seem like something a macroexpader should be doing. (Maybe
some kind of IR construct that converts any lambda leaf it recognizes
in a no arg checking policy instead? This could then be used in other
places, without the need to manaully pattern match for LAMBDA.)

Also, this allows us to get rid of COMPLEX-INITFORM stuff, since we
can just substitute the handler in directly. No need to create a new
function which calls the initform to be used as the actual
handler. Getting rid of this indirection probably makes things faster
too.