Count inline-expansions by component.
There are a few reasons for doing so:

Mainly, I have helped debug some projects (e.g. coalton) where the
compiler crashed compiling code due to heap exhaustion because too
much inlining was done in the component, with no reason being given or
any brake being triggered. This is because it is easy to construct
exponential blowup examples that do not actually count as recursive
calls, e.g. by having An inline An-1 twice, An-1 inline An-2 twice,
etc. down to A0. We don't want functions which seem to be "linear"
when considered in isolation to cause exponential-time blowup and
crash the compiler because that's a modularity violation that's proven
difficult to debug in real world projects.

The detection for recursive inlining is simpler - run-away recursive
inlining is a sufficient condition to trigger the component based
detector.

The only possible worry is that the arbitrary threshold prevents
inlining where it would otherwise be desirable - the note tells the
user suboptimality is happening, however, and the user can organize
functions so that the limit isn't exceeded or change the expansion
limit themselves, since *INLINE-EXPANSION-LIMIT* is exported.

Also, this makes the behavior of the code match the docstring for the
inline-expanded variable, since apparently it *did* used to be done
per component. This is especially important for an exported variable
in SB-EXT.