Fix quasi-infinite loop in arena allocator
A request for an object exceeding the extension block size would allocate a new
extension, fail to fit in that, again and again. Luckily the loop would terminate
(fatally) when the max extension count was reached. So use a discrete malloc
for oversized allocations. Also, even if a large request would fit in an extension
block, it could waste a large percentage of the previously active block.