Remove :LABELS keyword from ASSEMBLE macro - YAGNI.
Nothing used :LABELS until rev 8a19c68764 (HPPA support) which decided that
it had to. But HPPA is dead, and in no way was the keyword strictly needed,
nor do developers seem to use it. In a vop, it's just as well to see written
  (let ((x (gen-label))) ... (inst foo) ...)
versus
  (assemble (nil nil :labels (x)) ... (inst foo) ...)

And due to extreme obfuscation within that macro, its expansion was just plain
wrong in some cases, as per the included example.

Also this removes an (invisible) &OPTIONAL + &KEY style-warning.

If the newly-added style-warning about lexically shadowed labels
bothers some people, it's easy enough to take that out.