Make yet another enhancement to WITH-REGISTERS-PRESERVED
Allow it to avoid creating a new frame by specifying :frame-reg NIL.
In a non-leaf asm routine with a fast path and a slow path, you can link the
frame in to the call chain, save some args, test some conditions, and either
return quickly or call out. In the latter case, RBP should not be reassigned.
You have to be aware that RSP won't be restored after aligning it down to
save/restore FPRs when using the macro with the assumption that a frame was
already created.